Add an environment variable approach to disable saving a local file.
The environment variable is `RSERVE_DISABLE_LOCAL_FILE`.  If this
environment variable is set at compile time, the the `get_file` method
of the `Statistics::R::IO::Rserve` package will ignore the `$local`
parameter if set, and not save the file to that file name.  It will
merely return the `$data` retrieved from Rserve.

This will be used by WeBWorK and PG to prevent a problem from attempting
to save a file outside of the allowed html temporary directory.